Modularize E2EEnterYourPasswordStackNavigator
