+++ app/stacks/InsideStack/E2EEnterYourPasswordStackNavigator.js
@@ -1,0 +1,27 @@
+import React from "react";
+import { createStackNavigator } from "@react-navigation/stack";
+
+import E2EEnterYourPasswordView from "../views/E2EEnterYourPasswordView";
+
+const E2EEnterYourPasswordStack = createStackNavigator();
+const E2EEnterYourPasswordStackNavigator = () => {
+  const { theme } = React.useContext(ThemeContext);
+
+  return (
+    <E2EEnterYourPasswordStack.Navigator
+      screenOptions={{
+        ...defaultHeader,
+        ...themedHeader(theme),
+        ...StackAnimation,
+      }}
+    >
+      <E2EEnterYourPasswordStack.Screen
+        name="E2EEnterYourPasswordView"
+        component={E2EEnterYourPasswordView}
+        options={E2EEnterYourPasswordView.navigationOptions}
+      />
+    </E2EEnterYourPasswordStack.Navigator>
+  );
+};
+
+export default E2EEnterYourPasswordStackNavigator;
